requirements.txt
google-genai
Markdown
Pillow